requirements.txt
chromadb==0.4.20
SQLAlchemy==2.0.23
python-telegram-bot==20.7

arrow==1.3.0
# # linter 
# ipython
# pylint


# API Server
fastapi==0.108.0
pydantic==2.5.3
uvicorn==0.25.0
pyjwt==2.8.0
aiofiles==23.2.1
psutil==5.9.7

python-rapidjson==1.14
orjson==3.9.10
# semantic-kernel==0.3.12.dev0
semantic-kernel==0.4.5.dev0
# Notify systemd
sdnotify==0.3.2
GitPython==3.1.42
pyautogen==0.2.15
langchain==0.1.10

# docs 
sphinx-material
numpydoc
nbsphinx
recommonmark
sphinx-markdown-tables
sphinx_copybutton
readthedocs-sphinx-search
